+++ kytos/core/exceptions.py
@@ -112,17 +112,17 @@ class KytosSetTagRangeError(KytosTagError):
 
 
 class KytosTagsNotInTagRanges(KytosTagError):
-    """Exception thrown when a tag is outside of tag ranges"""
+    """Exception thrown when tags are outside of tag ranges"""
     def __init__(self, conflict: list[list[int]], intf_id: str) -> None:
-        msg = f"The tags {conflict} are outside tag_ranges in {intf_id}"
-        super().__init__(f"KytosSetTagRangeError, {msg}")
+        msg = f"The tags {conflict} are outside tag ranges in {intf_id}"
+        super().__init__(f"KytosTagsNotInTagRanges, {msg}")
 
 
 class KytosTagsAreNotAvailable(KytosTagError):
-    """Exception thrown when a tag is not available."""
+    """Exception thrown when tags are not available."""
     def __init__(self, conflict: list[list[int]], intf_id: str) -> None:
         msg = f"The tags {conflict} are not available in {intf_id}"
-        super().__init__(f"KytosSetTagRangeError, {msg}")
+        super().__init__(f"KytosTagsAreNotAvailable, {msg}")
 
 
 # Exceptions related  to NApps