+++ kytos/core/interface.py
@@ -23,8 +23,8 @@ from kytos.core.exceptions import (KytosSetTagRangeError,
 from kytos.core.helpers import now
 from kytos.core.id import InterfaceID
 from kytos.core.tag_ranges import (find_index_add, find_index_remove,
-                                   get_validated_tags, range_addition,
-                                   range_difference)
+                                   get_special_tags, get_validated_tags,
+                                   range_addition, range_difference)
 
 __all__ = ('Interface',)
 
@@ -155,8 +155,11 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
         self._tag_lock = Lock()
         self.available_tags = {'vlan': self.default_tag_values['vlan']}
         self.tag_ranges = {'vlan': self.default_tag_values['vlan']}
+        self.special_available_tags = {'vlan': self.default_special_tags}
+        self.special_tags = {'vlan': self.default_special_tags}
         self.set_available_tags_tag_ranges(
-            self.available_tags, self.tag_ranges
+            self.available_tags, self.tag_ranges,
+            self.special_available_tags, self.special_tags
         )
         super().__init__()
 
@@ -233,6 +236,12 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
         }
         return default_values
 
+    @property
+    def default_special_tags(self) -> list[str]:
+        """Reurn a default list of special tags. Applicable to
+         special_available_tags and special_tags."""
+        return ["untagged", "any"]
+
     def set_tag_ranges(self, tag_ranges: list[list[int]], tag_type: str):
         """Set new restriction, tag_ranges."""
         if tag_type != TAGType.VLAN.value:
@@ -269,6 +278,40 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
             )
             self.tag_ranges[tag_type] = self.default_tag_values[tag_type]
 
+    def set_special_tags(
+        self,
+        tag_type: str,
+        special_tags: list[str],
+    ):
+        """Set new restriction, special_tags"""
+        # Verify values in special_tags
+        tag_range = get_special_tags(
+            special_tags, self.default_special_tags
+        )
+
+        if tag_type != TAGType.VLAN.value:
+            msg = f"Tag type {tag_type} is not supported."
+            raise KytosTagtypeNotSupported(msg)
+        old_special_set = set(self.special_tags[tag_type])
+
+        for tag in self.special_available_tags[tag_type]:
+            old_special_set.remove(tag)  # Get special used tags
+        used_special = old_special_set.copy()
+
+        for tag in tag_range:
+            used_special.discard(tag)
+
+        # Missing used special used tags
+        if used_special:
+            msg = f"Missing tags in tag_range: {used_special}"
+            raise KytosSetTagRangeError(msg)
+
+        new_special_available = set(tag_range)
+        self.special_available_tags[tag_type] = list(
+            new_special_available - old_special_set
+        )
+        self.special_tags[tag_type] = tag_range
+
     def _remove_tags(self, tags: list[int], tag_type: str = 'vlan') -> bool:
         """Remove tags by resizing available_tags
         Returns False if nothing was remove, True otherwise"""
@@ -296,7 +339,7 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
     def use_tags(
         self,
         controller,
-        tags: Union[int, list[int], list[list[int]]],
+        tags: Union[str, int, list[int], list[list[int]]],
         tag_type: str = 'vlan',
         use_lock: bool = True,
         check_order: bool = True,
@@ -307,6 +350,7 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
         Args:
             controller: Kytos controller
             tags: value to be removed, multiple types for compatibility:
+                (str): Special vlan, "untagged" or "vlan"
                 (int): Single tag
                 (list[int]): Single range of tags
                 (list[list[int]]): List of ranges of tags
@@ -320,19 +364,27 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
         """
         if isinstance(tags, int):
             tags = [tags] * 2
-        elif check_order:
+        elif check_order and not isinstance(tags, str):
             tags = get_validated_tags(tags)
         if use_lock:
             with self._tag_lock:
-                self._use_tags(tags, tag_type)
+                try:
+                    self._use_tags(tags, tag_type)
+                except KeyError as err:
+                    msg = f"Tag type {err.args[0]} is not supported"
+                    raise KytosTagtypeNotSupported(msg)
         else:
-            self._use_tags(tags, tag_type)
+            try:
+                self._use_tags(tags, tag_type)
+            except KeyError as err:
+                msg = f"Tag type {err.args[0]} is not supported"
+                raise KytosTagtypeNotSupported(msg)
 
         self._notify_interface_tags(controller)
 
     def _use_tags(
         self,
-        tags: Union[list[int], list[list[int]]],
+        tags: Union[str, list[int], list[list[int]]],
         tag_type: str
     ):
         """Manage available_tags deletion changes."""
@@ -344,10 +396,15 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
                     self.available_tags[tag_type] = available_copy
                     conflict = range_difference(tags, available_copy)
                     raise KytosTagsAreNotAvailable(conflict, self._id)
-        else:
+        elif isinstance(tags, list):
             result = self._remove_tags(tags, tag_type)
             if result is False:
                 raise KytosTagsAreNotAvailable([tags], self._id)
+        else:
+            try:
+                self.special_available_tags[tag_type].remove(tags)
+            except ValueError:
+                raise KytosTagsAreNotAvailable(tags, self.id)
 
     # pylint: disable=too-many-branches
     def _add_tags(self, tags: list[int], tag_type: str = 'vlan') -> bool:
@@ -408,7 +465,7 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
     def make_tags_available(
         self,
         controller,
-        tags: Union[int, list[int], list[list[int]]],
+        tags: Union[str,  int, list[int], list[list[int]]],
         tag_type: str = 'vlan',
         use_lock: bool = True,
         check_order: bool = True,
@@ -418,6 +475,7 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
         Args:
             controller: Kytos controller
             tags: value to be added, multiple types for compatibility:
+                (str): Special vlan, "untagged" or "vlan"
                 (int): Single tag
                 (list[int]): Single range of tags
                 (list[list[int]]): List of ranges of tags
@@ -434,23 +492,31 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
         """
         if isinstance(tags, int):
             tags = [tags] * 2
-        elif check_order:
+        elif check_order and not isinstance(tags, str):
             tags = get_validated_tags(tags)
         if isinstance(tags[0], int) and tags[0] != tags[1]:
             tags = [tags]
         if use_lock:
             with self._tag_lock:
-                conflict = self._make_tags_available(tags, tag_type)
+                try:
+                    conflict = self._make_tags_available(tags, tag_type)
+                except KeyError as err:
+                    msg = f"Tag type {err.args[0]} is not supported"
+                    raise KytosTagtypeNotSupported(msg)
         else:
-            conflict = self._make_tags_available(tags, tag_type)
+            try:
+                conflict = self._make_tags_available(tags, tag_type)
+            except KeyError as err:
+                msg = f"Tag type {err.args[0]} is not supported"
+                raise KytosTagtypeNotSupported(msg)
         self._notify_interface_tags(controller)
         return conflict
 
     def _make_tags_available(
         self,
-        tags: Union[list[int], list[list[int]]],
+        tags: Union[str, list[int], list[list[int]]],
         tag_type: str,
-    ) -> list[list[int]]:
+    ) -> Union[str, list[list[int]], None]:
         """Manage available_tags adittion changes
 
         Exceptions:
@@ -464,15 +530,25 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
             new_tags, conflict = range_addition(tags, available_tags)
             self.available_tags[tag_type] = new_tags
             return conflict
-        result = self._add_tags(tags, tag_type)
-        if result is False:
-            return [tags]
-        return []
+        if isinstance(tags, list):
+            result = self._add_tags(tags, tag_type)
+            if result is False:
+                return [tags]
+        if isinstance(tags, str):
+            if tags not in self.special_tags[tag_type]:
+                raise KytosTagsNotInTagRanges(tags, self._id)
+            if tags not in self.special_available_tags[tag_type]:
+                self.special_available_tags[tag_type].append(tags)
+                return None
+            return tags
+        return None
 
     def set_available_tags_tag_ranges(
         self,
         available_tag: dict[str, list[list[int]]],
-        tag_ranges: dict[str, list[list[int]]]
+        tag_ranges: dict[str, list[list[int]]],
+        special_available_tags: dict[str, list[str]],
+        special_tags: dict[str, list[str]]
     ):
         """Set a range of VLAN tags to be used by this Interface.
 
@@ -483,6 +559,8 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
         with self._tag_lock:
             self.available_tags = available_tag
             self.tag_ranges = tag_ranges
+            self.special_available_tags = special_available_tags
+            self.special_tags = special_tags
 
     def enable(self):
         """Enable this interface instance.