+++ tests/unit/test_core/test_interface.py
@@ -157,19 +157,25 @@ class TestInterface():
         """Test available_tags and tag_ranges on Interface class."""
         default_available = {'vlan': [[1, 4095]]}
         default_tag_ranges = {'vlan': [[1, 4095]]}
-        intf_available = self.iface.available_tags
-        intf_tag_ranges = self.iface.tag_ranges
-        assert intf_available == default_available
-        assert intf_tag_ranges == default_tag_ranges
+        default_special_vlans = {'vlan': ["untagged", "any"]}
+        default_special_tags = {'vlan': ["untagged", "any"]}
+        assert self.iface.available_tags == default_available
+        assert self.iface.tag_ranges == default_tag_ranges
+        assert self.iface.special_available_tags == default_special_vlans
+        assert self.iface.special_tags == default_special_tags
 
         custom_available = {'vlan': [[10, 200], [210, 4095]]}
         custom_tag_ranges = {'vlan': [[1, 100], [200, 4095]]}
-        self.iface.set_available_tags_tag_ranges(custom_available,
-                                                 custom_tag_ranges)
-        intf_available = self.iface.available_tags
-        intf_tag_ranges = self.iface.tag_ranges
-        assert intf_available == custom_available
-        assert intf_tag_ranges == custom_tag_ranges
+        custom_special_vlans = {'vlan': ["any"]}
+        custom_special_tags = {'vlan': ["any"]}
+        self.iface.set_available_tags_tag_ranges(
+            custom_available, custom_tag_ranges,
+            custom_special_vlans, custom_special_tags
+        )
+        assert self.iface.available_tags == custom_available
+        assert self.iface.tag_ranges == custom_tag_ranges
+        assert self.iface.special_available_tags == custom_special_vlans
+        assert self.iface.special_tags == custom_special_tags
 
     async def test_interface_is_tag_available(self):
         """Test is_tag_available on Interface class."""
@@ -208,6 +214,9 @@ class TestInterface():
         with pytest.raises(KytosTagsAreNotAvailable):
             self.iface.use_tags(controller, [tags], use_lock=False)
         assert self.iface._notify_interface_tags.call_count == 3
+
+        self.iface.use_tags(controller, "untagged")
+        assert "untagged" not in self.iface.special_available_tags["vlan"]
 
     async def test_enable(self):
         """Test enable method."""
@@ -336,9 +345,12 @@ class TestInterface():
         """Test make_tags_available"""
         available = {'vlan': [[300, 3000]]}
         tag_ranges = {'vlan': [[20, 20], [200, 3000]]}
+        special_available_tags = {'vlan': []}
+        special_tags = {'vlan': ["any"]}
         self.iface._notify_interface_tags = MagicMock()
         self.iface.set_available_tags_tag_ranges(
-            available, tag_ranges
+            available, tag_ranges,
+            special_available_tags, special_tags
         )
         assert self.iface.available_tags == available
         assert self.iface.tag_ranges == tag_ranges
@@ -358,6 +370,13 @@ class TestInterface():
             self.iface.make_tags_available(controller, [1, 1], use_lock=False)
 
         assert self.iface.make_tags_available(controller, 300) == [[300, 300]]
+
+        with pytest.raises(KytosTagsNotInTagRanges):
+            self.iface.make_tags_available(controller, "untagged")
+
+        assert self.iface.make_tags_available(controller, "any") is None
+        assert "any" in self.iface.special_available_tags["vlan"]
+        assert self.iface.make_tags_available(controller, "any") == "any"
 
     async def test_set_tag_ranges(self, controller) -> None:
         """Test set_tag_ranges"""
@@ -388,12 +407,33 @@ class TestInterface():
         default = [[1, 4095]]
         assert self.iface.tag_ranges['vlan'] == default
 
+    def test_set_special_tags(self) -> None:
+        """Test set_special_tags"""
+        self.iface.special_available_tags["vlan"] = ["untagged"]
+        tag_type = "error"
+        special_tags = ["untagged", "any"]
+        with pytest.raises(KytosTagtypeNotSupported):
+            self.iface.set_special_tags(tag_type, special_tags)
+
+        tag_type = "vlan"
+        special_tags = ["untagged"]
+        with pytest.raises(KytosSetTagRangeError):
+            self.iface.set_special_tags(tag_type, special_tags)
+
+        special_tags = ["any"]
+        self.iface.set_special_tags(tag_type, special_tags)
+        assert self.iface.special_available_tags["vlan"] == []
+        assert self.iface.special_tags["vlan"] == ["any"]
+
     async def test_remove_tags(self) -> None:
         """Test _remove_tags"""
         available_tag = [[20, 20], [200, 3000]]
         tag_ranges = [[1, 4095]]
         self.iface.set_available_tags_tag_ranges(
-            {'vlan': available_tag}, {'vlan': tag_ranges}
+            {'vlan': available_tag},
+            {'vlan': tag_ranges},
+            {'vlan': ["untagged", "any"]},
+            {'vlan': ["untagged", "any"]}
         )
         ava_expected = [[20, 20], [241, 3000]]
         assert self.iface._remove_tags([200, 240])
@@ -415,7 +455,9 @@ class TestInterface():
         tag_ranges = [[1, 4095]]
         parameters = {
             "available_tag": {'vlan': available_tag},
-            "tag_ranges": {'vlan': tag_ranges}
+            "tag_ranges": {'vlan': tag_ranges},
+            "special_available_tags": {'vlan': ["untagged", "any"]},
+            "special_tags": {'vlan': ["untagged", "any"]}
         }
         self.iface.set_available_tags_tag_ranges(**parameters)
         assert self.iface._remove_tags([4, 6]) is False
@@ -427,7 +469,9 @@ class TestInterface():
         tag_ranges = [[1, 4095]]
         parameters = {
             "available_tag": {'vlan': available_tag},
-            "tag_ranges": {'vlan': tag_ranges}
+            "tag_ranges": {'vlan': tag_ranges},
+            "special_available_tags": {'vlan': ["untagged", "any"]},
+            "special_tags": {'vlan': ["untagged", "any"]}
         }
         self.iface.set_available_tags_tag_ranges(**parameters)
         ava_expected = [[4, 10], [20, 30]]
@@ -470,7 +514,9 @@ class TestInterface():
         tag_ranges = [[1, 4095]]
         parameters = {
             "available_tag": {'vlan': available_tag},
-            "tag_ranges": {'vlan': tag_ranges}
+            "tag_ranges": {'vlan': tag_ranges},
+            "special_available_tags": {'vlan': ["untagged", "any"]},
+            "special_tags": {'vlan': ["untagged", "any"]}
         }
         self.iface.set_available_tags_tag_ranges(**parameters)
         assert self.iface._add_tags([4, 6])